Feat - 삭제 기능 연결, isAuthor 값에 따라 메뉴 옵션 변경
